Added firebase functionality

+++ app/src/main/java/com/aniruddhapimple/portfolio/DataObjects.kt
@@ -7,6 +7,6 @@ data class ProjectDataObject(var title: String, var university : String, var dat
 
 data class EducationDataObject(var university: String, var location: String, var date: String, var otherDetails : String) : ListableItem
 
-data class SkillDataObject(var skill: String, var progress: Int)
+data class SkillDataObject(var skill: String, var progress: Long)
 
 interface ListableItem
